+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.0.7-dev+10.7.2013
 
-- Changed the `noClip` parameter of the `Anchor.setPosition` to a named optional
+- Changed the `noClip` parameter of `Anchor.setPosition` to a named optional
 parameter `clip` which defaults to `true`.
 
 ## 0.0.6+10.7.2013