+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.0.7-dev+10.7.2013
 
+- Added several public methods to the `Range` class with documentation and test 
+coverage.
 - Changed the `noClip` parameter of `Anchor.setPosition` to a named optional
 parameter `clip` which defaults to `true`.